refactored reporting tool testrail and tesults
