+++ page_objects/core_helpers/remote_objects.py
@@ -19,7 +19,7 @@ class Remote_Objects:
 
     def register_testrail(self):
         "Register TestRail with Page"
-        from utils.Test_Rail import Test_Rail  # pylint: disable=import-error,import-outside-toplevel
+        from integrations.reporting_tools.Test_Rail import Test_Rail  # pylint: disable=import-error,import-outside-toplevel
         self.testrail_flag = True
         self.testrail_object = Test_Rail()
         self.write('Automation registered with TestRail',level='debug')
@@ -35,7 +35,7 @@ class Remote_Objects:
     def register_tesults(self):
         "Register Tesults with Page"
         self.tesults_flag = True
-        from utils import Tesults # pylint: disable=import-error,import-outside-toplevel
+        from integrations.reporting_tools import Tesults # pylint: disable=import-error,import-outside-toplevel
         self.tesult_object = Tesults
 
     def report_to_testrail(self,case_id,test_run_id,result_flag,msg=''):
